Drops support for older versions of PHP and PHPUnit
