+++ .travis.yml
@@ -5,34 +5,6 @@ matrix:
     - php: nightly
   fast_finish: true
   include:
-    - php: 5.6
-      env:
-        - DEPS=lowest
-    - php: 5.6
-      env:
-        - DEPS=latest
-    - php: 5.6
-      env:
-        - PHPUNIT=minimum
-        - DEPS=latest
-    - php: 7.0
-      env:
-        - DEPS=lowest
-    - php: 7.0
-      env:
-        - DEPS=latest
-    - php: 7.1
-      env:
-        - DEPS=lowest
-    - php: 7.1
-      env:
-        - DEPS=latest
-    - php: 7.2
-      env:
-        - DEPS=lowest
-    - php: 7.2
-      env:
-        - DEPS=latest
     - php: 7.3
       env:
         - DEPS=lowest
@@ -55,27 +27,12 @@ matrix:
         - COMPOSER_FLAGS="--ignore-platform-reqs"
 
 install:
-  - if [[ $PHPUNIT == 'minimum' ]]; then sed -i 's/~5.7|/5.4.*|/g' ./composer.json ; fi
   - if [[ $DEPS == 'latest' ]]; then travis_retry composer update --no-interaction $COMPOSER_FLAGS ; fi
   - if [[ $DEPS == 'lowest' ]]; then travis_retry composer update --prefer-lowest --prefer-stable --no-interaction $COMPOSER_FLAGS ; fi
 
-before_script:
-  # Install extensions for PHP 5.x series. 7.x includes them by default.
-  - |
-    if [[ $TRAVIS_PHP_VERSION = 5.* ]]; then
-      cat <<< '
-        extension=mongo.so
-        extension=redis.so
-      ' >> ~/.phpenv/versions/"$(phpenv version-name)"/etc/conf.d/travis.ini
-    fi
-
 script:
 - |
-    if [[ $TRAVIS_PHP_VERSION = 5.* ]]; then
-      ./vendor/bin/phpunit --coverage-text --coverage-clover="build/logs/clover.xml" --testsuite="Mockery Test Suite PHP56";
-    else
-      ./vendor/bin/phpunit --coverage-text --coverage-clover="build/logs/clover.xml" --testsuite="Mockery Test Suite";
-    fi
+    ./vendor/bin/phpunit --coverage-text --coverage-clover="build/logs/clover.xml" --testsuite="Mockery Test Suite";
 
 after_success:
   - composer require satooshi/php-coveralls
@@ -100,5 +57,5 @@ deploy:
   github_token: $GITHUB_TOKEN
   on:
     branch: master
-    php: '7.1'
+    php: '7.3'
     condition: $DEPS = latest
