+++ library/Mockery/Adapter/Phpunit/TestListenerTrait.php
@@ -18,15 +18,7 @@
  * @license   http://github.com/padraic/mockery/blob/master/LICENSE New BSD License
  */
 
-namespace Mockery\Adapter\Phpunit\Legacy;
-
-if (class_exists('PHPUnit_Framework_TestCase') && ! class_exists('PHPUnit\Util\Blacklist')) {
-    class_alias('PHPUnit_Framework_ExpectationFailedException', 'PHPUnit\Framework\ExpectationFailedException');
-    class_alias('PHPUnit_Framework_Test', 'PHPUnit\Framework\Test');
-    class_alias('PHPUnit_Framework_TestCase', 'PHPUnit\Framework\TestCase');
-    class_alias('PHPUnit_Util_Blacklist', 'PHPUnit\Util\Blacklist');
-    class_alias('PHPUnit_Runner_BaseTestRunner', 'PHPUnit\Runner\BaseTestRunner');
-}
+namespace Mockery\Adapter\Phpunit;
 
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\Test;
