+++ tests/Mockery/Adapter/Phpunit/TestListenerTest.php
@@ -24,32 +24,12 @@ namespace tests\Mockery\Adapter\Phpunit;
 use Mockery\Adapter\Phpunit\MockeryTestCase;
 use PHPUnit\Framework\TestResult;
 use Mockery\Adapter\Phpunit\TestListener;
-
-if (class_exists('PHPUnit_Runner_Version') && version_compare(\PHPUnit_Runner_Version::id(), '6.0.0', '<')) {
-    class_alias('test\Mockery\Fixtures\EmptyTestCaseV5', 'tests\Mockery\Adapter\Phpunit\EmptyTestCase');
-} elseif (version_compare(\PHPUnit\Runner\Version::id(), '7.0.0', '<')) {
-    class_alias('test\Mockery\Fixtures\EmptyTestCaseV6', 'tests\Mockery\Adapter\Phpunit\EmptyTestCase');
-} else {
-    class_alias('test\Mockery\Fixtures\EmptyTestCaseV7', 'tests\Mockery\Adapter\Phpunit\EmptyTestCase');
-}
+use test\Mockery\Fixtures\EmptyTestCase;
 
 class TestListenerTest extends MockeryTestCase
 {
     protected function mockeryTestSetUp()
     {
-        /**
-         * Skip all tests here if PHPUnit is less than 6.0.0
-         */
-        if (class_exists('\PHPUnit\Runner\Version')) {
-            $ver = \PHPUnit\Runner\Version::series();
-        } else {
-            $ver = \PHPUnit_Runner_Version::series();
-        }
-
-        if (intval($ver) < 6) {
-            $this->markTestSkipped('The TestListener is only supported with PHPUnit 6+.');
-            return;
-        }
         // We intentionally test the static container here. That is what the
         // listener will check.
         $this->container = \Mockery::getContainer();
